+++ cryptocurrency-exchange/historicalrate.cpp
@@ -1,6 +1,6 @@
 #include "historicalrate.h"
 
-HistoricalRate::HistoricalRate(const tm& dt, const cryptoType& tCrypto,const float& cValue):
+HistoricalRate::HistoricalRate(const tm& dt, const cryptoType& tCrypto,const double& cValue):
     date(dt), cryptocurrency(tCrypto), value(cValue){}
 
 tm HistoricalRate::getDate()
@@ -8,7 +8,7 @@ tm HistoricalRate::getDate()
     return date;
 }
 
-float HistoricalRate::getValue()
+double HistoricalRate::getValue()
 {
     return value;
 }
